--- a/lib/galaxy/tools/__init__.py
+++ b/lib/galaxy/tools/__init__.py
@@ -2773,7 +2773,7 @@ class DatabaseOperationTool(Tool):
                 if not input_dataset_collection.collection.populated_optimized:
                     raise ToolInputsNotReadyException("An input collection is not populated.")
 
-            states, _  = input_dataset_collection.collection.dataset_states_and_extensions_summary
+            states, _ = input_dataset_collection.collection.dataset_states_and_extensions_summary
             for state in states:
                 check_dataset_state(state)