… lib/galaxy/tools/actions/model_operations.py
… lib/galaxy/tools/execute.py