For a collection of 1000 text files:

Before:
```
*** PROFILER RESULTS ***
create (/Users/mvandenb/src/galaxy/lib/galaxy/webapps/galaxy/api/tools.py:453)
function called 1 times

         60776745 function calls (59984285 primitive calls) in 59.550 seconds

   Ordered by: cumulative time, internal time, call count
   List reduced from 2068 to 40 due to restriction <40>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.004    0.004   59.665   59.665 tools.py:453(create)
        1    0.000    0.000   59.656   59.656 tools.py:468(_create)
        1    0.000    0.000   56.251   56.251 __init__.py:1577(handle_input)
        1    0.000    0.000   56.236   56.236 execute.py:29(execute)
        1    0.000    0.000   50.151   50.151 execute.py:49(execute_single_job)
        1    0.000    0.000   50.138   50.138 __init__.py:1631(handle_single_execution)
        1    0.002    0.002   50.138   50.138 __init__.py:1725(execute)
        1    0.000    0.000   50.135   50.135 model_operations.py:25(execute)
        1    0.002    0.002   40.479   40.479 model_operations.py:73(_produce_outputs)
        1    0.000    0.000   40.476   40.476 __init__.py:3156(produce_outputs)
     3014    0.007    0.000   36.452    0.012 scoping.py:162(do)
     1008    0.049    0.000   34.890    0.035 session.py:2489(flush)
     1005    0.034    0.000   34.825    0.035 session.py:2542(_flush)
        1    0.001    0.001   30.724   30.724 collections.py:487(apply_rules)
        1    0.008    0.008   30.716   30.716 collections.py:500(_build_elements_from_rule_data)
     1000    0.008    0.000   30.708    0.031 __init__.py:3161(copy_dataset)
     1000    0.007    0.000   27.115    0.027 tags.py:61(set_tags_from_list)
186191/167181    0.169    0.000   23.861    0.000 attributes.py:699(get)
221127/211125    0.147    0.000   22.083    0.000 attributes.py:279(__get__)
     1007    0.011    0.000   19.813    0.020 session.py:501(commit)
     1007    2.182    0.002   19.284    0.019 session.py:386(_remove_snapshot)
    10007    0.022    0.000   18.620    0.002 loading.py:190(load_on_ident)
    10007    0.071    0.000   18.598    0.002 loading.py:211(load_on_pk_identity)
    10009    0.009    0.000   18.494    0.002 query.py:3417(one)
    10009    0.206    0.000   18.485    0.002 query.py:3381(one_or_none)
    28036    0.041    0.000   18.255    0.001 base.py:952(execute)
    28036    0.033    0.000   18.204    0.001 elements.py:296(_execute_on_connection)
    28036    0.163    0.000   18.171    0.001 base.py:1088(_execute_clauseelement)
    22017    0.073    0.000   16.509    0.001 query.py:3501(_execute_and_instances)
  2528561    5.820    0.000   15.965    0.000 state.py:567(_expire)
    10010    0.016    0.000   14.692    0.001 query.py:3476(__iter__)
    30034    0.097    0.000   14.578    0.000 strategies.py:665(_load_for_state)
    12007    0.025    0.000   12.657    0.001 <string>:1(<lambda>)
    12007    0.131    0.000   12.631    0.001 strategies.py:772(_emit_lazyload)
    28036    0.230    0.000   12.237    0.000 base.py:1195(_execute_context)
    57983    0.035    0.000   11.129    0.000 state.py:640(_load_expired)
     6007    0.041    0.000   11.050    0.002 loading.py:938(load_scalar_attributes)
2580931/2571931    1.237    0.000   10.786    0.000 attr.py:257(__call__)
        2    0.002    0.001   10.727    5.363 __init__.py:246(_collect_inputs)
        5    0.000    0.000    9.337    1.867 __init__.py:21(visit_input_values)
```

After:
```
*** PROFILER RESULTS ***
create (/Users/mvandenb/src/galaxy/lib/galaxy/webapps/galaxy/api/tools.py:453)
function called 1 times

         9961471 function calls (9835656 primitive calls) in 12.098 seconds

   Ordered by: cumulative time, internal time, call count
   List reduced from 2082 to 40 due to restriction <40>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.005    0.005   12.148   12.148 tools.py:453(create)
        1    0.000    0.000   12.139   12.139 tools.py:468(_create)
        1    0.000    0.000    9.259    9.259 __init__.py:1576(handle_input)
        1    0.002    0.002    9.240    9.240 execute.py:29(execute)
153131/153128    0.083    0.000    7.009    0.000 attributes.py:279(__get__)
120174/109167    0.091    0.000    6.996    0.000 attributes.py:699(get)
        3    0.017    0.006    6.286    2.095 session.py:2489(flush)
        2    0.007    0.004    6.261    3.131 session.py:2542(_flush)
    13031    0.018    0.000    5.376    0.000 base.py:952(execute)
    13031    0.014    0.000    5.356    0.000 elements.py:296(_execute_on_connection)
    13031    0.066    0.000    5.342    0.000 base.py:1088(_execute_clauseelement)
    13031    0.098    0.000    4.451    0.000 base.py:1195(_execute_context)
     8016    0.024    0.000    3.719    0.000 query.py:3501(_execute_and_instances)
    22032    0.043    0.000    3.700    0.000 strategies.py:665(_load_for_state)
     5005    0.008    0.000    3.574    0.001 <string>:1(<lambda>)
     5005    0.044    0.000    3.566    0.001 strategies.py:772(_emit_lazyload)
        7    0.000    0.000    3.405    0.486 scoping.py:162(do)
        2    0.001    0.001    3.323    1.662 unitofwork.py:402(execute)
        1    0.000    0.000    3.017    3.017 __init__.py:1773(add_pending_datasets)
        1    0.000    0.000    3.017    3.017 __init__.py:1824(add_datasets)
     3006    0.005    0.000    3.010    0.001 loading.py:190(load_on_ident)
     3006    0.018    0.000    3.005    0.001 loading.py:211(load_on_pk_identity)
     3008    0.003    0.000    2.992    0.001 query.py:3417(one)
     3008    0.038    0.000    2.990    0.001 query.py:3381(one_or_none)
    13029    0.008    0.000    2.978    0.000 default.py:589(do_execute)
    13029    2.887    0.000    2.970    0.000 {method 'execute' of 'psycopg2.extensions.cursor' objects}
        1    0.002    0.002    2.967    2.967 __init__.py:1844(__add_datasets_optimized)
        1    0.000    0.000    2.917    2.917 mapping.py:2811(db_next_hid)
      3/2    0.000    0.000    2.905    1.452 session.py:899(begin)
      3/2    0.000    0.000    2.905    1.452 session.py:221(__init__)
      3/2    0.000    0.000    2.905    1.452 session.py:338(_take_snapshot)
        1    0.000    0.000    2.870    2.870 collections_util.py:106(dictify_dataset_collection_instance)
42595/36595    0.012    0.000    2.785    0.000 attr.py:257(__call__)
        1    0.001    0.001    2.774    2.774 collections_util.py:128(<listcomp>)
     1000    0.004    0.000    2.773    0.003 collections_util.py:168(dictify_element)
        2    0.002    0.001    2.763    1.382 base.py:59(before_flush)
     2000    0.055    0.000    2.758    0.001 __init__.py:3144(__create_version__)
       22    0.004    0.000    2.298    0.104 persistence.py:184(save_obj)
     3011    0.005    0.000    2.262    0.001 query.py:3476(__iter__)
        1    0.000    0.000    2.149    2.149 execute.py:49(execute_single_job)
```

The larger the collection the more significant the speedup.