@mvdbeek 2020-10-26
I guess we've been having reports about slow submission, so we could call it a performance bugfix. But we had the same reports for collection copies, history copies, history export etc. and putting all of them in last minute doesn't feel quite right. Maybe we can backport this after it has seen some usage on test.galaxyproject.org ?

I'm a little worried that maybe some weird corner case of tool submission for which we don't have a test is breaking now. Although we could still fix that ...